requirement.txt
absl-py==0.9.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==1.3.0
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.30
docker-pycreds==0.4.0
fasteners==0.17.3
filelock==3.8.0
fonttools==4.33.3
future==0.18.2
gfootball==2.9
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.3
greenlet==1.1.3
gtimer==1.0.0b5
gym==0.24.1
gym-notices==0.0.7
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.1
kiwisolver==1.4.3
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mujoco==2.2.0
mujoco-py==2.1.2.14
numpy==1.22.3
opencv-python==4.6.0.66
openseespy==3.4.0.2
openseespylinux==3.4.0.2
openseespymac==3.4.0.2
packaging==21.3
pandas==1.4.3
patchelf==0.14.5.0
pathtools==0.1.2
Pillow==9.0.1
pip==22.3.1
plotly==5.11.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycparser==2.21
pygame==2.1.2
pyglet==1.5.0
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rlkit==0.2.1.dev0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.10.1
sentry-sdk==1.6.0
setproctitle==1.2.3
setuptools==63.1.0
shortuuid==1.0.9
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
SQLAlchemy==1.4.41
tenacity==8.1.0
tensorboardX==2.0
threadpoolctl==2.2.0
tokenizers==0.13.2
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.20
wheel==0.37.1
zipp==3.8.0
